requirement.txt
google-genai>=1.51.0
langgraph>=0.2.39
streamlit>=1.40.0
python-dotenv
